Split codegen IR headers by concept
Move Function, Module/Global, and instruction/terminator types out of
basic_block.hpp into function.hpp, module.hpp, and instructions.hpp.
basic_block.hpp now only defines BasicBlock.
